upload_service/requirements.txt
fastapi
uvicorn
requests
python-multipart
mysql-connector-python
